Moved ProjectCarouselComponents to own folder
