+++ src/lib/components/ProjectCarousel/ProjectCard.svelte
