+++ src/lib/components/ProjectCarousel/ProjectCarousel.svelte
