+++ src/lib/components/ProjectCarousel/ProjectLinkCard.svelte
