+++ src/lib/components/ProjectCarousel/index.ts
@@ -1,0 +1,1 @@
+export { default as ProjectCarousel } from './ProjectCarousel.svelte';
